dbeaver/dbeaver-vscode#19 map custom folder as project in vc code product
